fix(linked-list): fixes the linked list
