+++ lerna.json
@@ -1,7 +1,7 @@
 {
   "npmClient": "yarn",
   "useWorkspaces": true,
-  "version": "0.0.2",
+  "version": "0.0.3",
   "packages": [
     "packages/*"
   ]
