+++ packages/linked-list/package.json
@@ -14,9 +14,6 @@
     "url": "git@github.com:yowainwright/datastructures-ts.git"
   },
   "keywords": [
-    "data-structures",
-    "datastructures",
-    "typescript",
     "linked list",
     "linkedlist",
     "linked-list"
