+++ packages/linked-list/src/__tests__/index.ts
@@ -1,8 +1,14 @@
 declare function test(msg: string, test: Function)
 declare function expect(result: any)
 
-import { LinkedList } from '..'
-import { Node } from '../Node'
+import { LinkedList, Node } from '..'
+
+test('Node init', function testNodeInit () {
+  const node = Node('foo')
+  expect(typeof node).toEqual('object')
+  expect(node.name).toEqual('foo')
+  expect(node.data).toEqual(null)
+})
 
 // test LinkedListNode(node)
 test('LinkedList init', function testLinkedList () {
